Heavy code refactoring, phase 1
- Robot environment consolidation
	* Created base class for quadruped robots
	* A1Walking, HoneyBadgerWalking, SilverBadgerWalking renamed
	  to A1Isaac, HoneyBadgerIsaac, SilverBadgerIsaac
	* reward weights tunable to avoid useless overrides
	* removed num'py from every file, except for the visualizer
- Import guard make importing isaac before launching the simulation app
  failing gracefully
- minor code refactoring an improvements, such as removing useless
  methods and converting others to static
- many methods moved to private
- Linting
- Added tests
