+++ docs/source/api/environments/isaacsim.rst
@@ -14,24 +14,31 @@ arrays and is only available once the simulation app is running. The app is star
 simulation backend, we rely on the Isaac Sim 6.0 default, PhysX; the launcher can select Newton instead.
 
 ``IsaacSim`` is the base class of every environment below. It owns the cloned scenes and builds the batched
-observation from the specification of the articulation entities to read.
+observation from the specification of the articulation entities to read. ``QuadrupedIsaac`` is a further base
+class, shared by the three legged robots below, holding the command-tracking logic and observations common to
+all of them; ``SilverBadgerIsaac`` in turn subclasses ``HoneyBadgerIsaac``, differing only in its robot-specific
+configuration.
 
 .. autosummary::
    :nosignatures:
 
    ~mushroom_rl.environments.isaacsim_env.IsaacSim
    ~mushroom_rl.environments.isaacsim_envs.cartpole.CartPoleIsaac
-   ~mushroom_rl.environments.isaacsim_envs.a1_walking.A1Walking
-   ~mushroom_rl.environments.isaacsim_envs.honey_badger_walking.HoneyBadgerWalking
-   ~mushroom_rl.environments.isaacsim_envs.silver_badger_walking.SilverBadgerWalking
+   ~mushroom_rl.environments.isaacsim_envs.quadruped.QuadrupedIsaac
+   ~mushroom_rl.environments.isaacsim_envs.a1.A1Isaac
+   ~mushroom_rl.environments.isaacsim_envs.honey_badger.HoneyBadgerIsaac
+   ~mushroom_rl.environments.isaacsim_envs.silver_badger.SilverBadgerIsaac
 
 .. automodule:: mushroom_rl.environments.isaacsim_env
     :private-members:
 
 .. automodule:: mushroom_rl.environments.isaacsim_envs.cartpole
 
-.. automodule:: mushroom_rl.environments.isaacsim_envs.a1_walking
+.. automodule:: mushroom_rl.environments.isaacsim_envs.quadruped
+    :private-members:
 
-.. automodule:: mushroom_rl.environments.isaacsim_envs.honey_badger_walking
+.. automodule:: mushroom_rl.environments.isaacsim_envs.a1
 
-.. automodule:: mushroom_rl.environments.isaacsim_envs.silver_badger_walking
+.. automodule:: mushroom_rl.environments.isaacsim_envs.honey_badger
+
+.. automodule:: mushroom_rl.environments.isaacsim_envs.silver_badger
