+++ mushroom_rl/environments/isaacsim_envs/__init__.py
@@ -1,6 +1,7 @@
 from .cartpole import CartPoleIsaac
-from .a1_walking import A1Walking
-from .honey_badger_walking import HoneyBadgerWalking
-from .silver_badger_walking import SilverBadgerWalking
+from .quadruped import QuadrupedIsaac
+from .a1 import A1Isaac
+from .honey_badger import HoneyBadgerIsaac
+from .silver_badger import SilverBadgerIsaac
 
-__all__ = ['CartPoleIsaac', 'A1Walking', 'HoneyBadgerWalking', 'SilverBadgerWalking']
+__all__ = ['CartPoleIsaac', 'QuadrupedIsaac', 'A1Isaac', 'HoneyBadgerIsaac', 'SilverBadgerIsaac']
