+++ mushroom_rl/utils/isaac_sim/_require_launched.py
@@ -1,0 +1,3 @@
+from .launcher import IsaacLauncher
+
+IsaacLauncher.require_running()
